Add files via upload
The idea for this paper is that it tries to improve the traditional momentum strategy by adding a panic state indicator when the market return demonstrates a divergent pattern in the short run and long run. It will then enter a long position for both stocks that performed well in the past and didn’t perform well. When the panic state indicator is not equal to 1, it will go back to the traditional momentum strategy, that is, enter a long position if the stock performed well in the past and enter a short position if the stock didn’t perform well in the past.

Just to mention that for simplicity I used the risk free rate == 0 when calculating the holding return when the strategy enters the panic state. Because it doesn’t matter when the strategy holds both long and short positions.